requirements.txt
tensorflow>=2.12.0
keras>=2.12.0
yfinance>=0.2.38
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
scikit-learn>=1.2.0
PyYAML>=6.0
tqdm>=4.65.0
